LayoutTests/media/video-dom-src.html

<video controls></video>
<script src=video-test.js></script>
<script>
test("video.currentSrc==''");
video.src = "content/test.mp4";
test("video.currentSrc==''");
waitForEventTestAndEnd("load", "relativeURL(video.currentSrc)=='content/test.mp4'");

video.load();
</script>
